Add files via upload
Flight Price Prediction is a machine learning-based approach to forecast the future prices of airline tickets. It involves analyzing historical data on flight prices, along with other factors such as:

- Time of year
- Demand
- Seasonality
- Competition
- Fuel prices
- Economic conditions

The goal is to build a model that can predict the future price of a flight with reasonable accuracy, helping:

- Airlines to optimize pricing strategies
- Travelers to make informed decisions about when to book flights
- Online travel agencies to provide personalized recommendations

By leveraging advanced algorithms and data analysis, flight price prediction can help stakeholders navigate the complex and dynamic aviation industry.